Attempt to fix crash on Linux

+++ Sources/GuideMEServer/Application/Model/DTO/Guide/Guide.swift
@@ -25,7 +25,7 @@ class Guide: Codable {
   convenience init(dict: [String: Any?]) {
     print("Decoding type")
     print("\(dict["type"])")
-    let type = Int(dict["type"] as! Int32)
+    let type = Int(dict["type"] as! Int64)
     var from: Int? = nil
     var to: Int? = nil
     print("Decoding from")
@@ -53,7 +53,7 @@ class Guide: Codable {
   }
 
   convenience init(dict: [String: Any?], city: City, preferenceType: [Int]) {
-    let type = Int(dict["type"] as! Int32)
+    let type = Int(dict["type"] as! Int64)
     var from: Int? = nil
     var to: Int? = nil
     if dict["from"] as? Int64 != nil {
